fix(layout): revert breakpoint removal
