--- packages/layout/src/spacing.js
@@ -1,48 +1,0 @@
-/**
- * Copyright IBM Corp. 2018, 2018
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import { miniUnits } from './mini-unit';
-
-// Spacing
-export const spacing01 = miniUnits(0.25);
-export const spacing02 = miniUnits(0.5);
-export const spacing03 = miniUnits(1);
-export const spacing04 = miniUnits(1.5);
-export const spacing05 = miniUnits(2);
-export const spacing06 = miniUnits(3);
-export const spacing07 = miniUnits(4);
-export const spacing08 = miniUnits(5);
-export const spacing09 = miniUnits(6);
-export const spacing = [
-  spacing01,
-  spacing02,
-  spacing03,
-  spacing04,
-  spacing05,
-  spacing06,
-  spacing07,
-  spacing08,
-  spacing09,
-];
-
-// Layout
-export const layout01 = miniUnits(2);
-export const layout02 = miniUnits(3);
-export const layout03 = miniUnits(4);
-export const layout04 = miniUnits(6);
-export const layout05 = miniUnits(8);
-export const layout06 = miniUnits(12);
-export const layout07 = miniUnits(20);
-export const layout = [
-  layout01,
-  layout02,
-  layout03,
-  layout04,
-  layout05,
-  layout06,
-  layout07,
-];
